cleaning up file structure for spectrometer_gui
