+++ src/qudi/config/spectrometer_testing.cfg
@@ -18,7 +18,7 @@ global:
 
 gui:
     spectrometer:
-        module.Class: 'spectrometer_gui.SpectrometerGui'
+        module.Class: 'spectrometer.spectrometer_gui.SpectrometerGui'
         connect:
             spectrumlogic: 'spectrumlogic'
